refact: 회원가입 관련 type 분리
