+++ src/components/signup/EmailForm.tsx
@@ -3,8 +3,8 @@ import Button from '@/components/common/Button/Button'
 import { Input } from '@/components/common/Input/Input'
 import { HelperLabel } from '@/components/common/HelperLabel/HelperLabel'
 import { Label } from '@/components/common/Label/Label'
-import { SignUpFormProps } from '@/app/(auth)/signup/page'
 import { sendEmail, verifyEmail } from '@/serverActions/auth/signup/actions'
+import { SignUpFormProps } from '@/types/signup/signupType'
 
 export const EmailForm = ({ next, setSignupForm }: SignUpFormProps) => {
   const [isCurrentStepValid, setIsCurrentStepValid] = useState(false)
